chore: typography components use their own theme which inherite from general theme
