+++ src/components/button/index.styles.js
@@ -1,7 +1,5 @@
 import styled from 'styled-components';
 
-import '../globalStyle.scss';
-
 export const ButtonContainer = styled.button`
   border: 1px solid blue;
   border-radius: ${({ theme, rounded }) => (rounded ? `${theme.button.borderRadius}px` : 0)};
