+++ src/components/index.js
@@ -3,6 +3,8 @@ import Button from './button';
 import Button1 from './button1';
 import Avatar from './avatar';
 
+import './globalStyle.scss';
+
 export { default as Typography } from './typography';
 export { default as Button } from './button';
 export { default as Button1 } from './button1';
