+++ src/components/themes.js
@@ -28,30 +28,6 @@ const defaultTheme = {
     riverBed: '#555f61',
     white: '#FFFFFF',
   },
-  typography: {
-    baseSize: 13,
-    lineHeightRatio: 1.75,
-    baseColor: '#3B4141',
-    h1SizeLarge: 30,
-    h1SizeMedium: 24,
-    h1SizeSmall: 18,
-    h2SizeLarge: 24,
-    h2SizeMedium: 20,
-    h2SizeSmall: 16,
-    h3SizeLarge: 20,
-    h3SizeMedium: 18,
-    h3SizeSmall: 14,
-    h4SizeLarge: 16,
-    h4SizeMedium: 14,
-    h4SizeSmall: 12,
-    h5SizeLarge: 14,
-    h5SizeMedium: 12,
-    h5SizeSmall: 10,
-    h6SizeLarge: 12,
-    h6SizeMedium: 10,
-    h6SizeSmall: 10,
-    headingLineHeightRatio: 1.68,
-  },
   sizes: {
     large: 'lg',
     medium: 'md',
