+++ src/components/typography/components/heading.styles.js
@@ -1,7 +1,6 @@
 import styled from 'styled-components';
 
 import { getSizeOfHeading, getMarginOfHeading } from '../utils';
-import '../../globalStyle.scss';
 
 export const H1 = styled.h1`
   font-family: 'IranSharp';
