+++ src/components/typography/components/link.test.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import 'jest-styled-components';
 
 import Link from './link';
-import { findByTestAtrr } from '../../helpers';
+import { findByTestAtrr } from '../../utils/testUtils';
 
 const setupShallow = (props = {}) => (
   shallow(
