+++ src/components/typography/components/paragraph.style.js
@@ -1,15 +1,9 @@
 import styled from 'styled-components';
 
-import '../../globalStyle.scss';
-
 export const Content = styled.p`
   margin: 8px 0;
   font-family: 'IranSharp';
-  font-size: ${({ theme, size }) => (size ? `${size}px` : `${theme.typography.baseSize}px`)};
-  line-height: ${({ theme, size }) => (
-    size
-      ? `${size * theme.typography.lineHeightRatio}px`
-      : `${theme.typography.baseSize * theme.typography.lineHeightRatio}px`
-  )};
-  color: ${({ theme }) => theme.typography.baseColor};
+  font-size: ${({ size }) => `${size}px`};
+  line-height: ${({ theme, size }) => `${size * theme.lineHeightRatio}px`};
+  color: ${({ theme }) => theme.defaultColor};
 `;
