+++ src/components/typography/index.js
@@ -1,19 +1,16 @@
 import Heading from './components/heading';
 import Text from './components/text';
 import Paragraph from './components/paragraph';
-import Title from './components/title';
 import Link from './components/link';
 
 export { default as Heading } from './components/heading';
 export { default as Text } from './components/text';
 export { default as Paragraph } from './components/paragraph';
-export { default as Title } from './components/title';
 export { default as Link } from './components/link';
 
 export default {
   Heading,
   Text,
   Paragraph,
-  Title,
   Link,
 };
