+++ src/components/utils/testUtils.js
@@ -1,0 +1,4 @@
+export const findByTestAtrr = (component, attr) => {
+  const wrapper = component.find(`[data-test='${attr}']`);
+  return wrapper;
+};
